修复 3 项代码审查发现的问题
1. gather_ppr_shards: 在 del edge_value 前保存其 dtype，
   修复空 shard 路径 UnboundLocalError

2. _csr_to_edge_index: 改为 int32 输出
   papers100M: 25.6GB(int64) → 12.8GB(int32)

3. _build_all_sub_edge_indices: 映射张量 int32 + 移除冗余 .long()
   node_to_part/global_to_local: 888MB→444MB 各
   edge_index 已是 int32，.long() 反而创建 12.8GB int64 拷贝

+++ NeutronGT/core/node_level_pipeline/graph_data.py
@@ -26,12 +26,13 @@ def load_optional_edge_csr(dataset_dir: str, dataset_name: str):
     return {"rowptr": rowptr, "col": col}
 
 def _csr_to_edge_index(edge_csr: dict):
-    rowptr = torch.as_tensor(edge_csr["rowptr"], dtype=torch.long, device="cpu")
-    col = torch.as_tensor(edge_csr["col"], dtype=torch.long, device="cpu")
+    # 使用 int32: papers100M 1.6B 边，int64 COO 25.6GB → int32 12.8GB
+    rowptr = torch.as_tensor(edge_csr["rowptr"], dtype=torch.int32, device="cpu")
+    col = torch.as_tensor(edge_csr["col"], dtype=torch.int32, device="cpu")
     if rowptr.numel() <= 1 or col.numel() == 0:
-        return torch.empty((2, 0), dtype=torch.long)
-    row_lengths = rowptr[1:] - rowptr[:-1]
-    rows = torch.arange(row_lengths.numel(), dtype=torch.long).repeat_interleave(row_lengths)
+        return torch.empty((2, 0), dtype=torch.int32)
+    row_lengths = rowptr[1:] - rowptr[:-1].to(torch.long)
+    rows = torch.arange(row_lengths.numel(), dtype=torch.int32).repeat_interleave(row_lengths)
     return torch.stack([rows, col], dim=0)
 
 def _get_node_degrees_from_csr(edge_csr: dict, num_nodes: int):
